Prepare refactor the Validator to be configurable and also to align to Single responsability
